% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesFPCA_package.R
\name{bayesFPCA-package}
\alias{bayesFPCA-package}
\title{bayesFPCA: an R package for FPCA by variational inference}
\description{
Suite of functions to perform functional principal component analysis (FPCA)
in the Bayesian setting using variational inference.
}
\details{
"_PACKAGE"
}
