% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cprod}
\alias{cprod}
\title{Cross-product between two vectors.}
\usage{
cprod(x, y)
}
\arguments{
\item{x}{Vector for the left hand side of the cross-product.}

\item{y}{Vector for the right hand side of the cross-product. If missing,
\code{x} is used.}
}
\value{
Resulting cross-product.
}
\description{
Function acting as  a wrapper on the R cross-product t(x) %*% y
}
