% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{display_eigenfunctions}
\alias{display_eigenfunctions}
\title{Display latent functions estimated by FPCA or mFPCA.}
\usage{
display_eigenfunctions(
  L,
  time_g,
  mu_g,
  Psi_g,
  mu_hat,
  list_Psi_hat,
  mu_hat_add = NULL,
  list_Psi_hat_add = NULL,
  mu_hat_ci = NULL,
  list_Psi_hat_ci = NULL,
  lwd = 2,
  data_col = "red",
  p_sample = NULL,
  vec_col_add = NULL,
  vec_lwd = NULL
)
}
\arguments{
\item{L}{Number of eigenfunctions (or latent dimensions).}

\item{time_g}{Vector for a dense time grid provided.}

\item{mu_g}{Reference mean function evaluated on the dense grid \code{time_g}.}

\item{Psi_g}{Reference eigenfunctions evaluated on the dense grid \code{time_g}.}

\item{mu_hat}{Estimated mean function as returned by
\code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}.}

\item{list_Psi_hat}{Estimated eigenfunctions as returned by
\code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}.}

\item{mu_hat_add}{Optional. Additional estimated mean function under the same
format as \code{mu_hat}.}

\item{list_Psi_hat_add}{Optional. Additional estimated mean function under the same
format as \code{list_Psi_hat_add}.}

\item{mu_hat_ci}{Optional. Credible boundaries for \code{mu_hat}.}

\item{list_Psi_hat_ci}{Optional. Credible boundaries for \code{list_Psi_hat}.}

\item{lwd}{Line width.}

\item{data_col}{Color for \code{mu_g} and \code{Psi_g}.}

\item{p_sample}{Indices of variables to display, if mFPCA. If FPCA, then will
be set to 1.}

\item{vec_col_add}{Vector of size 1 (if \code{mu_hat_add} and
\code{list_Psi_hat_add} are \code{NULL}) or 2 (if
\code{mu_hat_add} and \code{list_Psi_hat_add} are provided)
with the color(s) for \code{mu_hat} and
\code{list_Psi_hat} (first entry) and \code{mu_hat_add}
and \code{list_Psi_hat_add} (second entry).}

\item{vec_lwd}{Same as for \code{vec_col_add} but entries are line widths.}
}
\description{
This function is used to plot the mean function and eigenfunctions estimated
by \code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}, along with
reference functions.
}
\seealso{
\code{\link{flip_sign}} to flip the sign of eigenfunctions.
}
