% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{display_fit}
\alias{display_fit}
\title{Display univariate fits estimated by FPCA.}
\usage{
display_fit(N_sample, time_obs, time_g, Y, Y_hat, Y_low, Y_upp, offset = 0.1)
}
\arguments{
\item{N_sample}{Indices of samples to be displayed.}

\item{time_obs}{Vector containing time of observations for univariate curves.}

\item{time_g}{Vector for a dense time grid.}

\item{Y}{List of vectors with the functional data.}

\item{Y_hat}{List of estimated sample trajectories as returned by
\code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}.}

\item{Y_low}{List of estimated lower credible boundaries for the estimated
sample trajectories as returned by \code{\link{run_mfvb_fpca}}
or \code{\link{run_vmp_fpca}}.}

\item{Y_upp}{List of estimated upper credible boundaries for the estimated
sample trajectories as returned by \code{\link{run_mfvb_fpca}}
or \code{\link{run_vmp_fpca}}.}

\item{offset}{Optional offset to the plot y-axis.}
}
\description{
This function is used to plot the sample trajectories estimated from
univariate FPCA and returned by \code{\link{run_mfvb_fpca}} or
\code{\link{run_vmp_fpca}}.
}
\seealso{
\code{\link{display_fit_list}} for multivariate FPCA (mFPCA).
}
