% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{display_fit_list}
\alias{display_fit_list}
\title{Display multivariate fits estimated by mFPCA.}
\usage{
display_fit_list(
  p_sample,
  N_sample,
  time_obs,
  time_g,
  Y,
  Y_hat,
  Y_low,
  Y_upp,
  Y_hat_add = NULL,
  Y_low_add = NULL,
  Y_upp_add = NULL,
  offset = 0.1,
  col_data = "grey55",
  col = "black",
  col_add = "blue",
  lwd = 1.2,
  lwd_add = 1.2
)
}
\arguments{
\item{p_sample}{Indices of variables to be displayed.}

\item{N_sample}{Indices of samples to be displayed.}

\item{time_obs}{List of vectors containing time of observations for
multivariate curves.}

\item{time_g}{Vector for a dense time grid.}

\item{Y}{List of lists with the functional data.}

\item{Y_hat}{List of estimated sample trajectories as returned by
\code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}.}

\item{Y_low}{List of lower credible boundaries for the estimated sample
trajectories as returned by \code{\link{run_mfvb_fpca}}
or \code{\link{run_vmp_fpca}}.}

\item{Y_upp}{List of upper credible boundaries for the estimated sample
trajectories as returned by \code{\link{run_mfvb_fpca}}
or \code{\link{run_vmp_fpca}}.}

\item{Y_hat_add}{List of additional estimated sample trajectories under the
same format as returned by \code{\link{run_mfvb_fpca}} or
\code{\link{run_vmp_fpca}}.}

\item{Y_low_add}{List of additional lower credible boundaries for the estimated
sample trajectories under the same format as returned by
\code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}.}

\item{Y_upp_add}{List of additional upper credible boundaries for the estimated
sample trajectories under the same format as returned by
\code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}.}

\item{offset}{Optional offset to the plot y-axis.}

\item{col_data}{Color for the observed data points.}

\item{col}{Color for the estimated trajectories \code{Y_hat}.}

\item{col_add}{Color for the estimated trajectories \code{Y_hat_add}.}

\item{lwd}{Line width for the estimated trajectories \code{Y_hat}.}

\item{lwd_add}{Line width for the estimated trajectories \code{Y_hat_add}.}
}
\description{
This function is used to plot the sample trajectories from
multivariate FPCA (mFPCA) and returned by \code{\link{run_mfvb_fpca}} or
\code{\link{run_vmp_fpca}}
}
\seealso{
\code{\link{display_fit}} for univariate FPCA.
}
