% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{display_scores}
\alias{display_scores}
\title{Display the scores estimated by FPCA or mFPCA.}
\usage{
display_scores(
  N_sample,
  Zeta,
  Zeta_hat,
  zeta_ellipse,
  Zeta_hat_add = NULL,
  zeta_ellipse_add = NULL,
  vec_col = c("black", "blue"),
  data_col = "red",
  mfrow = NULL
)
}
\arguments{
\item{N_sample}{Indices of samples to be displayed.}

\item{Zeta}{Reference matrix of scores.}

\item{Zeta_hat}{Matrix of estimated scores as returned by
\code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}.}

\item{zeta_ellipse}{95\% posterior credible boundaries for the scores
\code{Zeta_hat} as returned by \code{\link{run_mfvb_fpca}}
or \code{\link{run_vmp_fpca}}.}

\item{Zeta_hat_add}{Optional. Additional matrix of estimated scores under the
same format as returned by \code{\link{run_mfvb_fpca}} or
\code{\link{run_vmp_fpca}}.}

\item{zeta_ellipse_add}{Optional. Additional 95\% posterior credible
boundaries for the scores \code{Zeta_hat} under the
same format as returned by \code{\link{run_mfvb_fpca}}
or \code{\link{run_vmp_fpca}}.}

\item{vec_col}{Vector of size 1 (if \code{Zeta_hat_add} is \code{NULL}) or 2 (if
\code{Zeta_hat_add} is provided) with the color(s) for
\code{Zeta_hat} (first entry) and \code{Zeta_hat_add}
(second entry).}

\item{data_col}{Color for \code{Zeta}.}

\item{mfrow}{Vector of integers of size two with the layout for plots. First
entry: number of rows, second entry: number of columns.}
}
\description{
This function is used to plot the scores estimated from
\code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}.
}
