% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{generate_fpca_data}
\alias{generate_fpca_data}
\title{Generate FPCA or mFPCA data.}
\usage{
generate_fpca_data(
  N,
  p,
  n,
  L,
  n_g,
  vec_sd_eps,
  mu_func,
  Psi_func,
  time_obs = NULL,
  format_univ = FALSE,
  generate_from_univ = FALSE,
  vec_sd_zeta = NULL,
  vec_rho_Zeta = NULL
)
}
\arguments{
\item{N}{Integer for the number of samples.}

\item{p}{Integer for the dimension of the curves (i.e., number of variables
measured longitudinally).}

\item{n}{Vector of size N for the number of time observations for each sample
(if univariate curves or multivariate curves with same number of
observations per variable) or matrix of size N x p where column j
contains the number of time observations for each sample for
variable j = 1, ..., p.}

\item{L}{Integer for the number of FPCA eigenfunctions (latent dimensions).}

\item{n_g}{Desired size of dense grid.}

\item{vec_sd_eps}{Vector of size p whose entry j contains the standard
deviation of the residual error for variable j = 1, ..., p.}

\item{mu_func}{Mean function, under the format return by
\code{\link{mu_func}}.}

\item{Psi_func}{Eigenfunctions, under the format return by
\code{\link{Psi_func}}.}

\item{time_obs}{Vector or list of vectors containing time of observations for
univariate or multivariate curves, respectively.}

\item{format_univ}{Boolean indicating whether the univariate format is used
in case of univariate curves.}

\item{generate_from_univ}{Boolean indicating whether to use the univariate
FPCA expansion to generate the curves. If
\code{FALSE}, the multivariate expansion is used
instead.}

\item{vec_sd_zeta}{Vector of size L whose entry l contains the standard
deviation of the scores for the latent dimension
l = 1, ..., L.}

\item{vec_rho_Zeta}{Vector of size p whose entry l contains the correlation
coefficient used to simulate correlated scores across
variables when the univariate expansion is used to
generate multivariate curves. Must be \code{NULL} if
\code{generate_from_univ} is \code{FALSE}.}
}
\value{
Object containing the generated time observations, dense grid,
        scores (Zeta), mean function, eigenfunctions and longitudinal curves
        (Y).
}
\description{
Function to generate univariate or multivariate functional curves based on
FPCA expansions.
}
