% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{Psi_func}
\alias{Psi_func}
\title{Example of two-dimensional latent function:
\code{( (-1)^j * sqrt(2/p)*cos(2*pi*t), (-1)^j * sqrt(2/p)*sin(2*pi*t) )}}
\usage{
Psi_func(t, j = 1, p = 1)
}
\arguments{
\item{t}{Time vector for which the function is evaluated.}

\item{j}{Sign swap; see definition of the function above.}

\item{p}{Parameter controlling the amplitude; see definition of the function
above.}
}
\value{
Function evaluated at \code{t}.
}
\description{
Example of two-dimensional latent function:
\code{( (-1)^j * sqrt(2/p)*cos(2*pi*t), (-1)^j * sqrt(2/p)*sin(2*pi*t) )}
}
