% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_algs.R
\name{run_mfvb_fpca}
\alias{run_mfvb_fpca}
\title{Mean-field variational Bayes (MFVB) inference for univariate or multivariate
functional principal component analysis.}
\usage{
run_mfvb_fpca(
  time_obs,
  Y,
  L,
  K = NULL,
  list_hyper = NULL,
  tol = 1e-05,
  maxit = 10000,
  rel_crit = TRUE,
  plot_elbo = FALSE,
  n_g = 1000,
  time_g = NULL,
  Psi_g = NULL,
  fixed_score_variance = TRUE,
  verbose = TRUE,
  seed = NULL,
  check_elbo = TRUE
)
}
\arguments{
\item{time_obs}{List of vectors or list of lists of vectors containing time
of observations for univariate or multivariate curves,
respectively.}

\item{Y}{List of vectors (for univariate curves) or list of lists of vectors
(for multivariate curves) with the functional data.}

\item{L}{Number of eigenfunctions (or latent dimensions).}

\item{K}{Number of O'Sulivan spline functions to be used in the FPCA or mFPCA
algorithms. If set to \code{NULL} will be set according to the rule
of Ruppert (2002), also enforcing K >=7.}

\item{list_hyper}{Hyperparameter settings constructed using the function
\code{\link{set_hyper}}. If \code{NULL} default hyperparameters will
be used.}

\item{tol}{Tolerance on the relative changes in the ELBO.}

\item{maxit}{Maximum number of iterations allowed.}

\item{rel_crit}{Boolean indicating whether the convergence criterion is
based on the relative (TRUE) or absolute (FALSE) changes in
the ELBO.}

\item{plot_elbo}{Boolean indicating whether the values of the ELBO should be displayed during the run.}

\item{n_g}{Desired size for dense grid.}

\item{time_g}{Dense grid provided as a vector of size \code{n_g}. If provided,
then \code{n_g} must be \code{NULL} as will be taken to be
\code{length(time_g)}.}

\item{Psi_g}{Reference eigenfunctions (if available, e.g., in simulations)
used to flip the sign of the resulting scores and eigenfunctions.}

\item{fixed_score_variance}{Boolean indicating whether to fix the variance of
the scores (to 1, estimating the variances of the latent function
spline coefficients only), or infer it in a component specific
way (along with the spline coefficient variances). Default TRUE.}

\item{verbose}{Boolean indicating whether messages should be printed during
the run.}

\item{seed}{User-specified seed for reproducibilty.}

\item{check_elbo}{If TRUE, checks that the ELBO (objective function) is increasing monotonically.}
}
\value{
An object containing the resulting MFVB estimates.
}
\description{
This function is used to perform FPCA or mFPCA using a MFVB algorithm.
}
\references{
Nolan, T. H., Goldsmith, J., & Ruppert, D. (2023). Bayesian Functional
Principal Components Analysis via Variational Message Passing with Multilevel
Extensions. Bayesian Analysis, 1(1), 1-27.

Ruppert, D. (2002). Selecting the number of knots for penalized splines.
Journal of computational and graphical statistics, 11(4), 735-757.
}
\seealso{
\code{\link{run_mfvb_fpca_model_choice}},
          \code{\link{run_vmp_fpca}}, \code{\link{run_vmp_fpca_model_choice}},
          \code{\link{set_hyper}}, \code{\link{flip_sign}}, \code{\link{display_fit}},
          \code{\link{display_fit_list}}, \code{\link{display_scores}},
          \code{\link{display_eigenfunctions}}
}
