% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_algs.R
\name{run_vmp_fpca}
\alias{run_vmp_fpca}
\title{Variational Message Passing (VMP) inference for univariate or multivariate
functional principal component analysis}
\usage{
run_vmp_fpca(
  time_obs,
  Y,
  L,
  K = NULL,
  list_hyper = NULL,
  n_g = 1000,
  time_g = NULL,
  tol = 1e-05,
  maxit = 10000,
  plot_elbo = FALSE,
  Psi_g = NULL,
  verbose = TRUE,
  seed = NULL
)
}
\arguments{
\item{time_obs}{Vector or list of vectors containing time of observations for
univariate or multivariate curves, respectively.}

\item{Y}{List of vectors (for univariate curves) or list of lists (for
multivariate curves) with the functional data.}

\item{L}{Number of eigenfunctions (or latent dimensions).}

\item{K}{Number of O'Sulivan spline functions to be used in the FPCA or mFPCA
algorithms. If set to \code{NULL} will be set according to the rule
of Ruppert (2002), also enforcing K >=7.}

\item{list_hyper}{Hyperparameter settings constructed using the function
\code{\link{set_hyper}}. If \code{NULL} default hyperparameters will
be used.}

\item{n_g}{Desired size for dense grid.}

\item{time_g}{Dense grid provided as a vector of size \code{n_g}. If provided,
then \code{n_g} must be \code{NULL} as will be taken to be
\code{length(time_g)}.}

\item{tol}{Tolerance on the relative changes in the ELBO.}

\item{maxit}{Maximum number of iterations allowed.}

\item{plot_elbo}{Boolean indicating whether the values of the ELBO should be
displayed during the run.}

\item{Psi_g}{Reference eigenfunctions (if available, e.g., in simulations)
used to flip the sign of the resulting scores and eigenfunctions.}

\item{verbose}{Boolean indicating whether messages should be printed during
the run.}

\item{seed}{User-specified seed for reproducibilty.}
}
\value{
An object containing the resulting VMP estimates.
}
\description{
This function is used to perform FPCA or mFPCA using a VMP algorithm.
}
\seealso{
\code{\link{run_mfvb_fpca}}, \code{\link{set_hyper}},
          \code{\link{flip_sign}}, \code{\link{display_fit}},
          \code{\link{display_fit_list}}, \code{\link{display_scores}},
          \code{\link{display_eigenfunctions}}
}
