% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{select_L}
\alias{select_L}
\title{Select the number of components L from the PVE estimate.}
\usage{
select_L(cumulated_pve, thres = 95, plot_pve = TRUE)
}
\arguments{
\item{cumulated_pve}{Vector of size L_max whose entry l contains the
PVE as estimated by the function bayesFPCA.}

\item{thres}{Optional threshold to be applied to the cumulated PVE to select
the number of components L (default 95).}

\item{plot_pve}{Boolean for whether the estimated cumulated PVE needs to be
displayed (default TRUE).}
}
\value{
An object containing the estimated number of components, based on the
        specified threshold on the cumulated PVE.
}
\description{
This function is used to select L from the estimated PVE (proportion of
variance explained) by the components and display the PVE as a function of
the component index l = 1, ..., L_max.
}
