% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyper.R
\name{set_hyper}
\alias{set_hyper}
\title{Gather model hyperparameters provided by the user.}
\usage{
set_hyper(sigma_zeta = 1, mu_beta = 0, sigma_beta = 1e+05, A = 1e+05)
}
\arguments{
\item{sigma_zeta}{Positive real number of the standard deviation of the
scores zeta.}

\item{mu_beta}{Vector of size 1 or 2 for the mean of the spline coefficients
for the mean function.}

\item{sigma_beta}{Vector of size 1 or 2 for the standard deviation of the
spline coefficients for the mean function.}

\item{A}{Positive real number for the top-level hyperparameter of the
iterated inverse chi-square distributions.}
}
\value{
An object containing the hyperparameter settings to be supplied to
        \code{\link{run_mfvb_fpca}} or \code{\link{run_vmp_fpca}}
}
\description{
This function is used to provide hyperparameter values for
\code{\link{run_mfvb_fpca}} and \code{\link{run_vmp_fpca}}.
}
\details{
The \code{\link{run_mfvb_fpca}} and \code{\link{run_vmp_fpca}} functions can
also be used with default hyperparameter values (i.e., without using
\code{\link{set_hyper}}) by setting their argument \code{list_hyper} to
\code{NULL}.
}
