% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{summarise_mcmc_multivariate}
\alias{summarise_mcmc_multivariate}
\title{Post-inference orthonormalisation procedure for MCMC inference [deprecated].}
\usage{
summarise_mcmc_multivariate(stan_obj, C_g, Psi_g)
}
\arguments{
\item{stan_obj}{Object obtained from R Stan code for FPCA inference.}

\item{C_g}{Design matrix C(t) constructed from the set of K spline functions
based on the dense time grid.}

\item{Psi_g}{Reference eigenfunctions (if available, e.g., in simulations)
used to flip the sign of the resulting scores and eigenfunctions.}
}
\value{
An object containing the orthnormalised eigenfunctions and
        uncorrelated scores.
}
\description{
This function is used to orthonormalise the eigenfunctions and scores
inferred using R Stan - deprecated, will be removed.
}
