% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesFPCA_package.R
\docType{package}
\name{bayesFPCA-package}
\alias{bayesFPCA-package}
\title{bayesFPCA: an R package for FPCA by variational message passing inference}
\description{
Suite of functions to perform functional principal component analysis (FPCA)
in the Bayesian setting using variational message passing inference.
}
\section{bayesFPCA functions}{
 gauss_mfpca_data, vmp_gauss_mfpca
}

