% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_algs.R
\name{run_mfvb_fpca}
\alias{run_mfvb_fpca}
\title{Mean-field variational Bayes (MFVB) inference for univariate or multivariate
functional principal component analysis.}
\usage{
run_mfvb_fpca(
  time_obs,
  Y,
  L,
  K = NULL,
  list_hyper = NULL,
  n_mfvb = 500,
  n_g = 1000,
  time_g = NULL,
  Psi_g = NULL,
  verbose = TRUE,
  seed = NULL
)
}
\arguments{
\item{time_obs}{Vector or list of vectors containing time of observations for
univariate or multivariate curves, respectively.}

\item{Y}{List of vectors (for univariate curves) or list of lists (for
multivariate curves) with the functional data.}

\item{L}{Number of eigenfunctions (or latent dimensions).}

\item{K}{Number of O'Sulivan spline functions to be used in the FPCA or mFPCA
algorithms. If set to \code{NULL} will be set according to the rule
of Ruppert (2002), also enforcing K >=7.}

\item{list_hyper}{Hyperparameter settings constructed using the function
\code{\link{set_hyper}}. If \code{NULL} default hyperparameters will
be used.}

\item{n_mfvb}{Number of iterations [ELBO and tolerance criterion to be
implemented soon].}

\item{n_g}{Desired size for dense grid.}

\item{time_g}{Dense grid provided as a vector of size \code{n_g}. If provided,
then \code{n_g} must be \code{NULL} as will be taken to be
\code{length(time_g)}.}

\item{Psi_g}{Reference eigenfunctions (if available, e.g., in simulations)
used to flip the sign of the resulting scores and eigenfunctions.}

\item{verbose}{Boolean indicating whether messages should be printed during
the run.}

\item{seed}{User-specified seed for reproducibilty.}
}
\value{
An object containing the resulting MFVB estimates.
}
\description{
This function is used to perform FPCA or mFPCA using a MFVB algorithm.
}
\references{
Nolan, T. H., Goldsmith, J., & Ruppert, D. (2023). Bayesian Functional
Principal Components Analysis via Variational Message Passing with Multilevel
Extensions. Bayesian Analysis, 1(1), 1-27.

Ruppert, D. (2002). Selecting the number of knots for penalized splines.
Journal of computational and graphical statistics, 11(4), 735-757.
}
\seealso{
\code{\link{run_vmp_fpca}}, \code{\link{set_hyper}},
          \code{\link{flip_sign}}, \code{\link{display_fit}},
          \code{\link{display_fit_list}}, \code{\link{display_scores}},
          \code{\link{display_eigenfunctions}}
}
