% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_algs.R
\name{run_mfvb_fpca_model_choice}
\alias{run_mfvb_fpca_model_choice}
\title{Mean-field variational Bayes (MFVB) inference for univariate or multivariate
functional principal component analysis with model choice procedure for
selection of the number of spline functions K.}
\usage{
run_mfvb_fpca_model_choice(
  time_obs,
  Y,
  L = 10,
  K_min = 5,
  K_max = 20,
  list_hyper = NULL,
  tol = 1e-05,
  maxit = 10000,
  rel_crit = TRUE,
  n_g = 1000,
  time_g = NULL,
  Psi_g = NULL,
  verbose = TRUE,
  seed = NULL,
  n_cpus = 1
)
}
\arguments{
\item{time_obs}{List of vectors or list of lists of vectors containing time
of observations for univariate or multivariate curves,
respectively.}

\item{Y}{List of vectors (for univariate curves) or list of lists of vectors
(for multivariate curves) with the functional data.}

\item{L}{Number of eigenfunctions (or latent dimensions).}

\item{K_min}{Minimum number of O'Sulivan spline functions K, which is modelled
             using an uniform prior distribution on the interval
.             [K_min, K_max]. Default K_min = 5.}

\item{K_max}{Maximum number of O'Sulivan spline functions K, which is modelled
             using an uniform prior distribution on the interval
.             [K_min, K_max]. Default K_max = 20.}

\item{list_hyper}{Hyperparameter settings constructed using the function
\code{\link{set_hyper}}. If \code{NULL} default hyperparameters will
be used.}

\item{tol}{Tolerance on the relative changes in the ELBO.}

\item{maxit}{Maximum number of iterations allowed.}

\item{rel_crit}{Boolean indicating whether the convergence criterion is
based on the relative (TRUE) or absolute (FALSE) changes in
the ELBO.}

\item{n_g}{Desired size for dense grid.}

\item{time_g}{Dense grid provided as a vector of size \code{n_g}. If provided,
then \code{n_g} must be \code{NULL} as will be taken to be
\code{length(time_g)}.}

\item{Psi_g}{Reference eigenfunctions (if available, e.g., in simulations)
used to flip the sign of the resulting scores and eigenfunctions.}

\item{verbose}{Boolean indicating whether messages should be printed during
the run.}

\item{seed}{User-specified seed for reproducibilty.}

\item{n_cpus}{Number of cores to be used for parallel runs. Set to 1 for serial run.}
}
\value{
An object containing the resulting MFVB estimates.
}
\description{
This function is used to perform FPCA or mFPCA using a MFVB algorithm.
}
\references{
Nolan, T. H., Goldsmith, J., & Ruppert, D. (2023). Bayesian Functional
Principal Components Analysis via Variational Message Passing with Multilevel
Extensions. Bayesian Analysis, 1(1), 1-27.

Ruppert, D. (2002). Selecting the number of knots for penalized splines.
Journal of computational and graphical statistics, 11(4), 735-757.
}
\seealso{
\code{\link{run_mfvb_fpca}},
          \code{\link{run_vmp_fpca}}, \code{\link{run_vmp_fpca_model_choice}},
          \code{\link{set_hyper}}, \code{\link{flip_sign}}, \code{\link{display_fit}},
          \code{\link{display_fit_list}}, \code{\link{display_scores}},
          \code{\link{display_eigenfunctions}}
}
